Extract enrichment, cleanup
